requirements.txt
flet==0.15.0
keyboard==0.13.5
librosa==0.10.1
matplotlib==3.8.0
Pillow==10.0.1
pydub==0.25.1
pygame==2.5.2
